Moved anim progress to separate folder with animations
