+++ src/game/graphics_handler/animations/tweener/animation_progress.js
